requirements.txt
# Web framework
Flask==2.3.3

# Deep Learning
torch>=2.0.0
transformers>=4.35.0

# Image processing
Pillow>=9.5.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization (training & evaluation)
matplotlib>=3.7.0

# Utilities
tqdm>=4.66.0
